backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML Models
transformers==4.35.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0

# Audio/Video Processing
librosa==0.10.1
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
scipy==1.11.4
moviepy==1.0.3
matplotlib==3.8.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# For SQLite (default)
# No additional package needed

# For PostgreSQL (optional)
# psycopg2-binary==2.9.9

# For MySQL (optional)
# pymysql==1.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# File Handling
aiofiles==23.2.1

# PDF Generation
reportlab==4.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2